fixup! PENDING: mtd: spinand: Call do_calibration when SPI NAND probes
This code is too TI specific (pattern size and position, operation to
perform). I would prefer something more generic in nand/spi/core.c.

Regarding the retrieval of the pattern, we actually do not need it to be
in the NAND in the first place, I guess we could have some kind of
command to write to the NAND cache and then read perform the PHY
calibration which reads from the NAND cache?