Enhance Winbond SPI NAND support
Both w25n**jw and w35n**jw chips have a "normal" and a "high speed"
mode. In order to use the high speed modes, we need to configure
internal registers and adapt the number of dummy cycles. The benefits
are too interesting for not paying attention to this little extra
configuration. In particular, it is an important building block for the
introduction of PHY calibration on TI SPI controllers. With these
changes combined, the frequency used on these chips can be bumped from
~25MHz up to 166MHz.

With this series applied, a `flash_speed -c10` on TI AM62A SK with a
W35N01JW shows:

	 eraseblock read speed is 9552 KiB/s
	 page read speed is 9516 KiB/s
	 2 page read speed is 9552 KiB/s

The extra modes do not bring a significant observable improvement
because the upstream support for this board does not yet allow the
highest supported frequency to be used. This is called PHY calibration,
and this is in the upstreaming process. With this extra work, it is
possible *much* better throughputs:

eraseblock read speed is 35555 KiB/s
page read speed is 33684 KiB/s
2 page read speed is 35068 KiB/s

Link: https://github.com/miquelraynal/linux/tree/winbond/6.16-rc1/octal-phy

In order to introduce all these variants and derive the quickest one, I
had to improve a bit the helper deriving the time an ops would
take. These changes can go through the spi tree, the other patches do
not depend on them and the performance hit is rather acceptable without.

While at adding maximum operation frequencies, I realized I got myself
confused with the macro parameters due to some of them being
optional (with variable arguments in macros). I decided it was too error
prone so I propose to add these values on all READ_FROM_CACHE
variants (where they are often relevant).

To: Mark Brown <broonie@kernel.org>
To: Richard Weinberger <richard@nod.at>
To: Vignesh Raghavendra <vigneshr@ti.com>
Cc: Yogesh S <yogeshs@ti.com>
Cc: Santhosh Kumar K <s-k6@ti.com>
Cc: Steam Lin <STLin2@winbond.com>
Cc: Thomas Petazzoni <thomas.petazzoni@bootlin.com>
Cc: linux-spi@vger.kernel.org
Cc: linux-kernel@vger.kernel.org
Cc: linux-mtd@lists.infradead.org
Signed-off-by: Miquel Raynal <miquel.raynal@bootlin.com>

--- b4-submit-tracking ---
# This section is used internally by b4 prep for tracking purposes.
{
  "series": {
    "revision": 1,
    "change-id": "20250613-winbond-6-16-rc1-octal-phy-upstream-c8929842ed21",
    "prefixes": []
  }
}